Abstract most variables into global scope for ui
This is mainly because a lot of functions need these in a lot of scopes.
Later on these will be in their own class, so will be available as if
they are in global scope.

This allows for easy update/references of the selected entry, the
resources and the user input.
Split resources in original and filtered variants, and use filtered for
almost all cases, so that things like references and selection are
relevant for the displayed list and not one that just lies in memory.