Add first draft of an App class
This should handle all the basic app functionality like storing the
windows, updating them, handling the user input etc etc.

This first version still has a mix of things in the main.cpp and the
app.cpp. This will be changed over time until everything is better
handled in the app.cpp.