Decouple filtering and displaying a bit
Before both things were done with the internal knowledge of the other.
Meaning that filter needed to give display all of the internal structure
(hidden flag, things to hide even tho they are in the vector etc.).
This should be slowly removed over time.

This is the first step, the collection just returns a vector which has
the filtered items inside, meaning the display does not need to care
about other items which are hidden.
One thing that still needs internal knowledge is to get the index of the
selected entry, because if we do not pass a translation from full vector
to the filtered one, the display does not know which one to show and
more importantly, the collection does not know which one was chosen.