Remove non blocking way of iteration
For now I am not using the non blocking, later on I might want or need
it. The reason for now is, that when it does not block, it runs on all
resources, meaning it uses 100% of one of my cores.

This needs to be addressed with a sleep or anything else later on. But
for now can be ignored.