Add basic config json validation
Instead of crashing the application if the user adds a wrong config or
misses a value, the app will now display all missing configs in an
output and not crash.

It still does not handle general exceptions from the parser (like when
the json format is off [missing "," etc.]) but this should not be as
often as misspelling an option.